fix nil pointer due to init execution order
Golang executes multiple `init()` within a package in filename order. This
meant that the `init()` in "registry.go", which initialises the `registry`
was being called *after* the `init()` in "dev_info_collector.go", which
was attempting to add an item to that registry. The result was a nil pointer
exception.

This commit applies a quick fix by adding a new `RegisterCollector()` function
that initialises `registry` on the first call.

However, this is not a complete solution as the registry is currently part of
the `collectors` package, and therefore could still be accessed before being
properly initialised. A full solution will move the registry into its own
package to allow use of private and public interfaces to ensure this is
not possible.

Signed-off-by: Charlie Wheeler-Robinson <cwheeler@redhat.com>